+++ src/jk_prettyprintobj/__init__.py
@@ -1,11 +1,10 @@
 ﻿
 
-__version__ = "0.2021.2.10"
+__version__ = "0.2021.2.17"
 
 
 
-from .dumper import DumpMixin, Dumper, DumpCtx
-
+from .dumper import DumpMixin, Dumper, DumpCtx, DEFAULT_DUMPER_SETTINGS