+++ HISTORY.md
@@ -3,6 +3,10 @@
 	* Provided full documentation of the module
 
 * 2021-02-10
-
 	* Added empty __slots__ to DumpMixin to prevent classes derived from getting a __dict__
+
+* 2021-02-17
+	* Improved _canCompactSequence()
+	* Corrected type in dumper.py
+	* Improved value type detection